use float instead of doubles for correlation coefficients

+++ Correlation.cpp
@@ -5,21 +5,21 @@
 // Computes the correlation coefficient of two rows of image data "x" and "y" with a length of "length" bytes.
 // The correlation coefficient is defined as the result of dividing the empiric covariance by the empiric variance.
 
-double CorrelationCoefficient(unsigned char* x, unsigned char* y, int length) {
+float CorrelationCoefficient(unsigned char* x, unsigned char* y, int length) {
 	// calculate means of x and y
-	double xSum = 0.0;
-	double ySum = 0.0;
+	float xSum = 0.0f;
+	float ySum = 0.0f;
 	for (int i = 0; i < length; i++) {
 		xSum += x[i];
 		ySum += y[i];
 	}
-	double xMean = xSum / length;
-	double yMean = ySum / length;
+	float xMean = xSum / length;
+	float yMean = ySum / length;
 	
-	double covariance = 0, varianceX = 0, varianceY = 0;
+	float covariance = 0, varianceX = 0, varianceY = 0;
 	for (int i = 0; i < length; i++) {
-		double xDiff = x[i] - xMean;
-		double yDiff = y[i] - yMean;
+		float xDiff = x[i] - xMean;
+		float yDiff = y[i] - yMean;
 		
 		covariance += xDiff * yDiff;
 		varianceX += xDiff * xDiff;
@@ -27,7 +27,7 @@ double CorrelationCoefficient(unsigned char* x, unsigned char* y, int length) {
 	}
 	
 	// prevent division by zero
-	if (varianceX * varianceY < DBL_EPSILON) {
+	if (varianceX * varianceY < FLT_EPSILON) {
 		return 0;
 	}
